[v0.91.4][maintainability] Split governed executor by responsibility
